Always recurse to process followees of people we follow (#5)
Under normal circumstances SSB clients build a list of feeds which should be
replicated by analysing contact messages. Each contact messages creates a
directed edge connecting two vertices constituting relevant feeds in the
so-called social graph. To create a list of feeds to replicate the graph is
walked following the directed edges up to a certain depth, starting with the
local node.

This behaviour is more complicated in go-ssb. The library will only follow the
edge if the edge is mutual (sometimes referred to as "friends"). This was
reported as counterintuitive as other clients do not behave in this way.

This commit changes this behaviour to the default behaviour described in the
first paragraph of this commit message.